control_plane/requirements.txt
Django==5.1.8
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
redis==5.0.1
whitenoise==6.9.0
